refactor tags, questions actions
