+++ src/actions/queries.js
@@ -4,7 +4,7 @@ import queryApi from "../api/query";
 export const getLines = () => (dispatch, getState) => {
   const { filters } = getState().filtersWrapper || {};
   dispatch({ type: types.SEARCH_QUERY_REQUEST });
-  queryApi
+  return queryApi
     .sendQuery(filters)
     .then(res => {
       dispatch({ type: types.SEARCH_QUERY_REQUEST_SUCCESS, data: res });
@@ -22,9 +22,9 @@ export const searchQueryRequest = ({ query }) => {
 };
 
 export const searchQueryFromTagRequest = ({ query }) => {
-  return function (dispatch) {
+  return async function (dispatch) {
     dispatch({ type: types.SEARCH_QUERY_FROM_TAG_REQUEST, query });
-    dispatch(getLines());
+    return dispatch(getLines());
   };
 };
 
@@ -38,12 +38,8 @@ export const getCurrentQueryId = state => {
   return state.linesWrapper.queryId || [];
 };
 
-export const onTagPress = ({ tag }) => {
-  return function (dispatch) {
-    dispatch({ type: types.ON_TAG_CLICK, tag });
-    dispatch({ type: types.SEARCH_QUERY_FROM_TAG_REQUEST, query: tag });
-    dispatch(getLines());
-  };
+export const getCurrentLines = state => {
+  return state.linesWrapper.linesList || [];
 };
 
 export const showLikedLines = () => {
@@ -82,25 +78,6 @@ export const changePage = ({ pageChangeTo }) => {
   return function (dispatch) {
     dispatch({ type: types.FILTER_PAGE_CHANGE, data: pageChangeTo });
     dispatch(getLines());
-  };
-};
-
-async function getQuestions(lines) {
-  let questions = lines.map(line => queryApi.getQuestion({ query: line }));
-  questions = await Promise.all(questions);
-  return questions;
-}
-
-export const generateQuestions = ({ lines }) => {
-  return async function (dispatch) {
-    dispatch({ type: types.ASQ_QUESTIONS_REQUEST });
-    getQuestions(lines)
-      .then(response => {
-        dispatch({ type: types.ASQ_QUESTIONS_SUCCESS, data: response });
-      })
-      .catch(err => {
-        return dispatch({ type: types.ASQ_QUESTIONS_FAILURE, error: err });
-      });
   };
 };
